requirements.txt
python-telegram-bot>=20.0
zxcvbn>=4.4.28
nltk>=3.6.0
python-dotenv>=0.19.0
